+++ Gui/MyRequests.py
@@ -43,15 +43,26 @@ class MyRequests(QMainWindow, ui_MyRequests.Ui_MainWindow):
                 QTreeWidgetItem([i[0], i[3], i[1], i[2], i[5]]))
 
     def __DeleteSentRequest(self):
-        DBConnection.execute(RequestTable.delete().where(
-            RequestTable.c.Title == self.SentRequests.selectedItems()[0].text(0),
-            RequestTable.c.Username == self.SentRequests.selectedItems()[0].text(1),
-            RequestTable.c.Details == self.SentRequests.selectedItems()[0].text(2),
-            RequestTable.c.Price == self.SentRequests.selectedItems()[0].text(3),
-        ))
-        self.__refresh()
+        if len(self.SentRequests.selectedItems()) == 0:
+            errDlg = ErrorDialog("Please select a request!")
+            errDlg.exec()
+            return
+        self.dlg = CDialog(
+            "Are you sure you want to delete this request?", "Question!", self)
+        if self.dlg.exec():
+            DBConnection.execute(RequestTable.delete().where(
+                RequestTable.c.Title == self.SentRequests.selectedItems()[0].text(0),
+                RequestTable.c.Username == self.SentRequests.selectedItems()[0].text(1),
+                RequestTable.c.Details == self.SentRequests.selectedItems()[0].text(2),
+                RequestTable.c.Price == self.SentRequests.selectedItems()[0].text(3),
+            ))
+            self.__refresh()
 
     def __ViewSentRequest(self):
+        if len(self.SentRequests.selectedItems()) == 0:
+            errDlg = ErrorDialog("Please select a request!")
+            errDlg.exec()
+            return
         self.ViewReqWnd = ViewRequest.ViewRequest(DBConnection.execute(RequestTable.select().where(
             RequestTable.c.Title == self.SentRequests.selectedItems()[0].text(0),
             RequestTable.c.Username == self.SentRequests.selectedItems()[0].text(1),
@@ -60,7 +71,11 @@ class MyRequests(QMainWindow, ui_MyRequests.Ui_MainWindow):
         )).fetchall()[0][6])
 
     def __ViewReceivedRequest(self):
-            self.ViewReqWnd = ViewRequest.ViewRequest(DBConnection.execute(RequestTable.select().where(
+        if len(self.ReceivedRequests.selectedItems()) == 0:
+            errDlg = ErrorDialog("Please select a request!")
+            errDlg.exec()
+            return
+        self.ViewReqWnd = ViewRequest.ViewRequest(DBConnection.execute(RequestTable.select().where(
             RequestTable.c.Title == self.ReceivedRequests.selectedItems()[0].text(0),
             RequestTable.c.Username == self.ReceivedRequests.selectedItems()[0].text(1),
             RequestTable.c.Details == self.ReceivedRequests.selectedItems()[0].text(2),
@@ -68,23 +83,33 @@ class MyRequests(QMainWindow, ui_MyRequests.Ui_MainWindow):
         )).fetchall()[0][6])
 
     def __AcceptReceivedRequest(self):
-        DBConnection.execute(RequestTable.update().where(
-            RequestTable.c.Title == self.ReceivedRequests.selectedItems()[0].text(0),
-            RequestTable.c.Username == self.ReceivedRequests.selectedItems()[0].text(1),
-            RequestTable.c.Details == self.ReceivedRequests.selectedItems()[0].text(2),
-            RequestTable.c.Price == self.ReceivedRequests.selectedItems()[0].text(3),
-        ).values(Status = "Waiting for admin"))
-        self.__refresh()
+        if len(self.ReceivedRequests.selectedItems()) == 0:
+            errDlg = ErrorDialog("Please select a request!")
+            errDlg.exec()
+            return
+        self.dlg = CDialog(
+            "Are you sure you want to accept this request?", "Question!", self)
+        if self.dlg.exec():
+            DBConnection.execute(RequestTable.update().where(
+                RequestTable.c.Title == self.ReceivedRequests.selectedItems()[0].text(0),
+                RequestTable.c.Username == self.ReceivedRequests.selectedItems()[0].text(1),
+                RequestTable.c.Details == self.ReceivedRequests.selectedItems()[0].text(2),
+                RequestTable.c.Price == self.ReceivedRequests.selectedItems()[0].text(3),
+            ).values(Status = "Waiting for admin"))
+            self.__refresh()
 
     def __DeclineReceivedRequest(self):
         if len(self.ReceivedRequests.selectedItems()) == 0:
             errDlg = ErrorDialog("Please select a request!")
             errDlg.exec()
             return
-        DBConnection.execute(RequestTable.delete().where(
-            RequestTable.c.Title == self.ReceivedRequests.selectedItems()[0].text(0),
-            RequestTable.c.Username == self.ReceivedRequests.selectedItems()[0].text(1),
-            RequestTable.c.Details == self.ReceivedRequests.selectedItems()[0].text(2),
-            RequestTable.c.Price == self.ReceivedRequests.selectedItems()[0].text(3),
-        ))
-        self.__refresh()
+        self.dlg = CDialog(
+            "Are you sure you want to decline this request?", "Question!", self)
+        if self.dlg.exec():
+            DBConnection.execute(RequestTable.delete().where(
+                RequestTable.c.Title == self.ReceivedRequests.selectedItems()[0].text(0),
+                RequestTable.c.Username == self.ReceivedRequests.selectedItems()[0].text(1),
+                RequestTable.c.Details == self.ReceivedRequests.selectedItems()[0].text(2),
+                RequestTable.c.Price == self.ReceivedRequests.selectedItems()[0].text(3),
+            ))
+            self.__refresh()
